+++ src/api_template/core/models.py
@@ -5,9 +5,9 @@ from sqlmodel import Field, SQLModel
 
 
 class Model(SQLModel):
-    id: UUID = Field(uuid4(), primary_key=True)
-    created_at: datetime = Field(datetime.now(UTC), nullable=False)
-    updated_at: datetime = Field(datetime.now(UTC), nullable=False)
+    id: UUID = Field(default_factory=uuid4, primary_key=True)
+    created_at: datetime = Field(default_factory=lambda: datetime.now(UTC), nullable=False)
+    updated_at: datetime = Field(default_factory=lambda: datetime.now(UTC), nullable=False)
 
 
 class UserModel(Model, table=True):