+++ src/api_template/use_cases/auth.py
@@ -21,7 +21,7 @@ class AuthUseCase:
         if self.user_repository.read_by_username(username) is not None:
             raise UsernameAlreadyExistsError
         user = self.user_repository.create(username, self.hash_service.hash(password))
-        return self.jwt_service.encode({"id": str(user.id)}, settings.jwt_key, settings.jwt_expires_in_seconds)
+        return self.jwt_service.encode({"sub": str(user.id)}, settings.jwt_key, settings.jwt_expires_in_seconds)
 
     def sign_in(self, username: str, password: str) -> str:
         user = self.user_repository.read_by_username(username)
@@ -29,11 +29,11 @@ class AuthUseCase:
             raise BadCredentialsError
         if not self.hash_service.check(password, user.password):
             raise BadCredentialsError
-        return self.jwt_service.encode({"id": str(user.id)}, settings.jwt_key, settings.jwt_expires_in_seconds)
+        return self.jwt_service.encode({"sub": str(user.id)}, settings.jwt_key, settings.jwt_expires_in_seconds)
 
     def get_user(self, token: str) -> User:
         payload = self.jwt_service.decode(token, settings.jwt_key)
-        user = self.user_repository.read_by_id(UUID(payload["id"]))
+        user = self.user_repository.read_by_id(UUID(payload["sub"]))
         if user is None:
             raise UserDoesNotExistError
         return user