+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changed
 
+- 🗃️ Change models default values
 - 🎨 Increase line width
 
 ## [1.0.0] - 2025-07-08